% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{acns_date_added}
\alias{acns_date_added}
\title{Pull Latest \code{date_added} from ACNS Data}
\usage{
acns_date_added(.data)
}
\arguments{
\item{.data}{ACNS data with a \code{date_added} column}
}
\value{
The latest \code{date_added} in \code{Date} format, or \code{NA_Date_} if not
available
}
\description{
Pull Latest \code{date_added} from ACNS Data
}
