% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acns_date_updated.R
\name{acns_date_updated}
\alias{acns_date_updated}
\title{Get Modification Time of ACNS File from SFTP Server}
\usage{
acns_date_updated(
  usr = Sys.getenv("sftp_usr"),
  pwd = Sys.getenv("sftp_pwd"),
  path = "ACNS/ACNS_DAILY_OUT.txt"
)
}
\arguments{
\item{usr}{Username for SFTP server}

\item{pwd}{Password for SFTP server}

\item{regexp}{Regular expression identifying the ACNS file}
}
\value{
A \code{Date}
}
\description{
Get Modification Time of ACNS File from SFTP Server
}
