% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_acns_upload.R
\name{archive_acns_upload}
\alias{archive_acns_upload}
\title{Archive Upload Data for ACNS System}
\usage{
archive_acns_upload(
  .data,
  dir = "V:/EPI DATA ANALYTICS TEAM/Mass Texting/ACNS_DAILY_OUT/Combined/",
  date = attr(.data, "date")
)
}
\arguments{
\item{.data}{Data from \code{translate_acns_upload()}}

\item{dir}{The archive directory}

\item{date}{The date on the filename}
}
\value{
\code{.data} (invisibly)
}
\description{
Archive Upload Data for ACNS System
}
