% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_positive.R
\name{archive_positive}
\alias{archive_positive}
\title{Archive ACNS Upload}
\usage{
archive_positive(.data, dir, date = attr(.data, "date"))
}
\arguments{
\item{.data}{A data frame or data fram extension (e.g. a \code{tibble})}

\item{dir}{The archive directory}

\item{date}{The date to use as a timestamp}
}
\value{
\code{.data} (invisibly)
}
\description{
\code{archive_positive()} creates a backup of the ACNS data upload.
}
