% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_positive.R
\name{as_single_digit_date}
\alias{as_single_digit_date}
\title{Convert \code{Date} Objects to "M/D/YYYY" Format}
\usage{
as_single_digit_date(x)
}
\arguments{
\item{x}{A \code{Date} vector}
}
\value{
A \code{character} vector of formatted dates
}
\description{
\code{as_single_digit_date()} converts \code{Date} vectors to "M/D/YYYY" \code{character}
format, where month and day do not contain leading zeros (i.e. April 1, 2020
would be 4/1/2020, not 04/01/2020).
}
