% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_acns.R
\name{bind_acns_positive}
\alias{bind_acns_positive}
\title{Bind Positive NBS Data to ACNS}
\usage{
bind_acns_positive(.acns, assign = FALSE, date = attr(.acns, "date"), ...)
}
\arguments{
\item{.acns}{ACNS data}

\item{assign}{Is this call for case assignment purposes (\code{TRUE}) or
SMS purposes (\code{FALSE})? Default is \code{FALSE}.}

\item{date}{The \code{date} attribute for the \code{date_tbl} output}

\item{...}{Additional arguments to pass to
\code{\link[covidsms:prep_positive]{prep_positive()}}; at time of creation,
these include \code{filter_lab} and \code{filter_acns}}
}
\description{
\code{bind_acns_positive()} binds new data from \code{prep_positive()} to the input
ACNS data.
}
