% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_new_sms.R
\name{check_new_sms}
\alias{check_new_sms}
\title{Check for New ACNS SMS File (ACNS_DAILY_OUT)}
\usage{
check_new_sms()
}
\value{
\code{logical}
}
\description{
\code{check_new_sms()} downloads the SMS file on the SFTP server and compares it
to the latest local SMS file. If these two are not identical, it returns
\code{TRUE}; otherwise, \code{FALSE}.
}
