% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{class-date_tbl}
\alias{class-date_tbl}
\alias{as_date_tbl}
\alias{is_date_tbl}
\alias{new_date_tbl}
\alias{validate_date_tbl}
\title{\code{date_tbl}: A \code{tibble} with a \code{date} Attribute}
\usage{
as_date_tbl(
  x,
  date,
  .rows = NULL,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

is_date_tbl(x)

new_date_tbl(x, date, nrow = vctrs::vec_size(x))

validate_date_tbl(x)
}
\arguments{
\item{x}{A \code{date_tbl}, \code{tibble}-like object, or an object to test for
inheritance from \code{date_tbl}. In \code{as_date_tbl()}, this can be anything
coercable to a \code{tibble}; \code{new_date_tbl()} does less work and is
somewhat stricter about the input. \code{is_date_tbl()} and
\code{validate_date_tbl()} will take any object, but will return \code{FALSE} or an
error (respectively) if the object is not a properly created \code{date_tbl}.}

\item{date}{A \code{Date} or datetime (\code{POSIXt}) object of length 1}

\item{.rows}{The number of rows, useful to create a 0-column tibble or
just as an additional check.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{rownames}{How to treat existing row names of a data frame or matrix:
\itemize{
\item \code{NULL}: remove row names. This is the default.
\item \code{NA}: keep row names.
\item A string: the name of a new column. Existing rownames are transferred
into this column and the \code{row.names} attribute is deleted.
Read more in \link[tibble]{rownames}.
}}

\item{nrow}{The number of rows, required.}
}
\value{
A \code{date_tbl}
}
\description{
\code{date_tbl()} is a subclass of \code{tibble} that stores a \code{date} attribute as
metadata. This is useful, for example, in tracking the report date of a
dataset.

\code{as_date_tbl()} converts an object to a \code{date_tbl}

\code{is_date_tbl()} tests whether an object is a \code{date_tbl}

\code{new_date_tbl()} constructs a \code{date_tbl} from a \code{tibble}-like object and a
date with minimal checking

\code{validate_date_tbl()} validates the properties of a \code{date_tbl}
}
\keyword{internal}
