% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-paths.R
\docType{data}
\name{data-paths}
\alias{data-paths}
\alias{ltcf_addr_path}
\alias{ltcf_phone_path}
\title{Paths to Data on COVID-19 ("V:") Drive}
\format{
An object of class \code{character} of length 1.

An object of class \code{fs_path} (inherits from \code{character}) of length 1.
}
\usage{
ltcf_addr_path

ltcf_phone_path
}
\description{
These variables contain paths to datasets housed on the Shelby County
COVID-19 ("V:") drive.

\code{ltcf_addr_path} returns the path to an \code{fst} file containing parsed
address information for long-term care facilities

\code{ltcf_phone_path} returns the path to an \code{fst} file containing parsed phone
numbers for long-term care facilities
}
\keyword{datasets}
