% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialr.R
\name{detect_java}
\alias{detect_java}
\title{Detect Java Install Using Environment Variables}
\usage{
detect_java(dialr = FALSE, quiet = FALSE)
}
\arguments{
\item{dialr}{Should warning messages be formatted for
\code{\link[covidsms:test_dialr]{test_dialr()}}?}

\item{quiet}{Should warning messages be displayed at all?}
}
\value{
A boolean indicating whether Java is available
}
\description{
\code{detect_java()} attempts to check that the JAVA_HOME variable is set and that
the PATH includes a JDK bin. If either of these things is not true, it
returns \code{FALSE}.
}
\keyword{internal}
