% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_investigation.R
\name{distinct_investigation}
\alias{distinct_investigation}
\title{Deduplicate Investigations}
\usage{
distinct_investigation(.data)
}
\arguments{
\item{.data}{Input data from \code{translate_positive()}}
}
\value{
Deduplicated data
}
\description{
\code{distinct_investigations()} coalesces duplicate investigations down to a
single record.
}
\details{
This function is memoised; it will cache return values and load them rather
than re-processing the same input data.
}
