% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{download_acns}
\alias{download_acns}
\title{Download Current ACNS Data}
\usage{
download_acns(
  primary = c("sms", "addr"),
  addr_creds = Sys.getenv(c("acns_usr", "acns_pwd")),
  sms_creds = Sys.getenv(c("sftp_usr", "sftp_pwd"))
)
}
\arguments{
\item{primary}{Which data source should be used as the primary table? Choices
are \code{"sms"} or \code{"addr"}; defaults to \code{"sms"}.}

\item{addr_creds}{Character. A length 2 vector containing the credentials for
accessing the ACNS address data. The username should appear first, then the
password.}

\item{sms_creds}{Character. A length 2 vector containing the credentials for
accessing the ACNS SMS data. The username should appear first, then the
password.}
}
\value{
A \code{tibble} containing the joined data
}
\description{
\code{download_acns()} downloads data from both the address and texting ACNS
files and merges the two using common columns. It uses the address data as
reference; any data in the SMS dataset not matching the address data is
dropped. This is because address data is necessary for some downstream
functions.
}
