% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{download_sftp}
\alias{download_sftp}
\title{Download File from SFTP Server}
\usage{
download_sftp(
  file,
  usr,
  pwd,
  dir_remote = "",
  dir_local = fs::file_temp("dir"),
  server = "xfer.shelbycountytn.gov"
)
}
\arguments{
\item{file}{File name to pull from the SFTP server. This should
not include directories.}

\item{usr}{Username for SFTP server}

\item{pwd}{Password for SFTP server}

\item{dir_remote}{Directory containing file in SFTP server}

\item{dir_local}{Directory to save file to on local computer}

\item{server}{Server address}
}
\value{
Path where the file is downloaded
}
\description{
Download File from SFTP Server
}
