% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{download_sms}
\alias{download_sms}
\title{Download ACNS_DAILY_OUT File from SFTP Server}
\usage{
download_sms(
  path = "ACNS/ACNS_DAILY_OUT.txt",
  usr = Sys.getenv("sftp_usr"),
  pwd = Sys.getenv("sftp_pwd"),
  server = "xfer.shelbycountytn.gov",
  archive = TRUE,
  archive_dir = "V:/EPI DATA ANALYTICS TEAM/Mass Texting/ACNS_DAILY_OUT/",
  ...
)
}
\arguments{
\item{path}{Path to file to download on SFTP server}

\item{usr}{Username for SFTP server}

\item{pwd}{Password for SFTP server}

\item{server}{Server address}

\item{archive}{Should the data be archived before loading in-memory? This
will only write data to an archive file that does not already exist.}

\item{archive_dir}{The archive directory path}

\item{...}{Additional arguments to pass to \code{read_file()}}
}
\value{
A \code{tibble} containing the data from the ACNS sms file
}
\description{
\code{download_sms()} downloads the file used to send text messages. Date columns
are returned as dates; all other are character. All column names are passed
through \code{janitor::clean_names()}.
}
