% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_residence.R
\name{filter_by_residence}
\alias{filter_by_residence}
\title{Filter Positive Records by Jurisdiction/County/State Residence}
\usage{
filter_by_residence(.data)
}
\arguments{
\item{.data}{The NBS data to filter; should be the result of joining a
PCR and investigations file}
}
\description{
\code{filter_by_residence()} filters out non-Shelby-County residents, while
retaining individuals for whom residence is verified or unsure.
}
