% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialr.R
\name{install_java}
\alias{install_java}
\title{Install OpenJDK for Windows}
\usage{
install_java(dir = "~", check_java = TRUE)
}
\arguments{
\item{dir}{Directory to save the JDK folder}

\item{check_java}{Should the function attempt to detect Java prior to
installation?}
}
\value{
The path to the OpenJDK directory (invisibly)
}
\description{
\code{install_java()} downloads and makes available the current release version
of OpenJDK, which is required for the dialr package. It sets the \code{JAVA_HOME}
variable and adds the JDK "bin" folder to the user \code{PATH}.
}
\keyword{internal}
