% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chr_ascii.R
\name{is_chr_ascii}
\alias{is_chr_ascii}
\alias{is_chr_ascii.data.frame}
\alias{is_chr_ascii.default}
\title{Test that an Object is ACSCII Character Encoded}
\usage{
is_chr_ascii(x, na_rm = FALSE)

\method{is_chr_ascii}{data.frame}(x, na_rm = FALSE)

\method{is_chr_ascii}{default}(x, na_rm = FALSE)
}
\arguments{
\item{x}{An object to test}

\item{na_rm}{Should missing values be removed for testing? If \code{FALSE}, any
missing values will return \code{NA}}
}
\value{
\code{TRUE} if the object is ASCII character, \code{NA} if it includes missing
values and \code{na_rm = FALSE}, otherwise \code{FALSE}
}
\description{
\code{is_chr_ascii()} tests whether an atomic vector or data frame is entirely an
ASCII-encoded character.
}
