% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_positive.R
\name{load_nbs_positive}
\alias{load_nbs_positive}
\alias{load_inv_positive}
\alias{load_pcr_positive}
\title{Load Positive Tests or Cases From NBS Snapshot File}
\usage{
load_pcr_positive(date = NULL)

load_inv_positive(date = NULL)
}
\arguments{
\item{date}{Optional. Character (formatted as "YYYY-MM-DD") or \code{Date}.}
}
\description{
\code{load_pcr_positive()} loads positive tests from the latest PCR snapshot file,
or the one specified by \code{date}, if provided. \code{load_inv_positive()} does the
same for the latest investigations snapshot file.
}
