% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_positive.R
\name{load_new_positive}
\alias{load_new_positive}
\title{Load New NBS Positive PCRs for a Given Date}
\usage{
load_new_positive(date = NULL)
}
\arguments{
\item{date}{Optional. Character (formatted as "YYYY-MM-DD") or \code{Date}.}
}
\value{
A \code{date_tbl}, which is a \code{tibble} with a \code{date} attribute
}
\description{
\code{load_new_positive()} loads newly reported positive tests on a given date.
It identifies newly reported tests on that date, subsets to those tests, and
adds information from the investigations file to the identified records.
}
