% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_positive.R
\name{load_positive}
\alias{load_positive}
\title{Load PCR Positives from a Given Date with Investigations File Data}
\usage{
load_positive(date = NULL)
}
\arguments{
\item{date}{Optional. Character (formatted as "YYYY-MM-DD") or \code{Date}.}
}
\value{
A \code{date_tbl}, which is a \code{tibble} with a \code{date} attribute
}
\description{
\code{load_positive()} loads positive PCR tests and investigations, and it joins
the results together.
}
