% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoised.R
\name{mem_if_not}
\alias{mem_if_not}
\title{Memoise a Function if Not Already Done}
\usage{
mem_if_not(f, ...)
}
\arguments{
\item{f}{The function to memoise}

\item{...}{Additional arguments to pass to \code{memoise()}}
}
\value{
The memoised function
}
\description{
Memoise a Function if Not Already Done
}
