% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{no_fmt}
\alias{no_fmt}
\title{Standardize vroom/readr \code{col_spec}s with No Format}
\usage{
no_fmt(col_spec)
}
\arguments{
\item{col_spec}{A column specification from vroom/readr}
}
\value{
The column specification with format set to \code{""}
}
\description{
Standardize vroom/readr \code{col_spec}s with No Format
}
