% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ltcf_addr1.R
\name{parse_ltcf_addr1}
\alias{parse_ltcf_addr1}
\title{Parse \code{ADDR1} Field to House Number + Street Name}
\usage{
parse_ltcf_addr1(.data, .col = "ADDR1")
}
\arguments{
\item{.data}{A data frame or data frame extension}

\item{.col}{\verb{<tidy-select>} The column containing address information. This
must match only one column, which must be character}
}
\value{
\code{.data} with the added column \code{.addr1_tmp_} containing the parsed
address string
}
\description{
\code{parse_ltcf_addr1()} parses out house numbers and street names (excluding
directionals and suffixes) from the \code{ADDR1} field in positive/ACNS data.
}
