% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_sms.R
\name{path_sms}
\alias{path_sms}
\title{Path to SMS File}
\usage{
path_sms(
  date = NULL,
  dir = "V:/EPI DATA ANALYTICS TEAM/Mass Texting/ACNS_DAILY_OUT/",
  combined = FALSE
)
}
\arguments{
\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{dir}{The directory holding the files of interest}

\item{combined}{Should the combined ACNS/NBS file be returned? If \code{TRUE},
appends \code{"Combined"} to the end of the directory path.}
}
\value{
An \code{fs_path} character vector
}
\description{
Path to SMS File
}
