% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_positive.R
\name{pkey_nbs}
\alias{pkey_nbs}
\title{Create Primary Key from an NBS ID}
\usage{
pkey_nbs(string)
}
\arguments{
\item{string}{A character vector of NBS IDs}
}
\value{
A character vector of primary keys
}
\description{
\code{pkey_nbs()} creates a primary key of the format "{NBS}-{ID_NUMBER}" from
an 8-digit unique identifier in NBS. For labs, this is the \code{lab_local_id};
for investigations, this is \code{inv_local_id}; for people, this \emph{should} be
\code{patient_local_id}, though many people have multiple \code{patient_local_id}s.
}
