% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_acns.R
\name{prep_acns}
\alias{prep_acns}
\title{Prepare ACNS Data for Use in Case Assignment}
\usage{
prep_acns(.data = download_acns(), date = attr(.data, "date"))
}
\arguments{
\item{.data}{Data frame or data frame extension containing ACNS data. If none
is supplied, the latest data will be downloaded.}

\item{date}{The date that the data was reported. This is extracted from the
\code{date} attribute of a \code{date_tbl}.}
}
\value{
Prepared data in \code{tibble} format
}
\description{
\code{prep_acns()} prepares ACNS data for use in case assignment and NBS data
management processes. In addition to standardizing the dataset, it adds 3
variables:
\describe{
\item{school_age}{A logical indicating whether a case is school-aged}
\item{long_term_care}{A logical indicating whether a case is housed in a
long-term care facility}
\item{duplicate}{A logical indicating whether a case is a duplicate}
}
}
