% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_positive.R
\name{prep_positive}
\alias{prep_positive}
\title{Transform NBS Data to Format Necessary for Exporting to ACNS File}
\usage{
prep_positive(
  .data = load_positive(),
  filter_lab = TRUE,
  labs = c("AEL", "BAPTIST", "CCHS", "POPLAR", "UT"),
  filter_new = FALSE,
  filter_acns = filter_new,
  date = attr(.data, "date")
)
}
\arguments{
\item{.data}{Data from NBS snapshot file}

\item{filter_lab}{Should records from labs reported through the ACNS process
be filtered out?}

\item{filter_new}{Should only new positives be returned? These are still
de-duplicated across the full NBS dataset.}

\item{filter_acns}{Should duplicates from previous ACNS files be removed?}

\item{date}{The \code{date} attribute for the resulting \code{date_tbl}}
}
\value{
A \code{date_tbl}, which is a \code{tibble} with a \code{date} attribute
}
\description{
\code{prep_positive()} transforms data from NBS snapshot files to the
format needed for exports to the ACNS process.
}
