% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoised.R
\name{read-memoised}
\alias{read-memoised}
\alias{read_inv_positive}
\alias{read_pcr_positive}
\title{Memoised \code{read_file_delim()} Wrappers}
\usage{
read_inv_positive(path, ...)

read_pcr_positive(path, ...)
}
\arguments{
\item{path}{The path to the file to read}

\item{...}{Additional arguments to pass to \code{read_file_delim()}}
}
\value{
A \code{tibble} containing the loaded data
}
\description{
These memoised wrappers are used to speed up repeated loading of the same
files. They only re-load the file if an argument has changed. This allows
functions like \code{load_inv_positive()} to run \emph{much} more quickly when
subsequent calls with the same arguments are made.
}
