% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{read_file}
\alias{read_file}
\title{Read Delimited/Fixed-Width/Excel File}
\usage{
read_file(
  file,
  col_types = vroom::cols(.default = vroom::col_character()),
  ...
)
}
\arguments{
\item{file}{File to read}

\item{col_types}{readr/vroom column specification}

\item{...}{Additional arguments to pass to \code{vroom()} or \code{read_excel()}}
}
\value{
A \code{tibble} containing the file data
}
\description{
Read Delimited/Fixed-Width/Excel File
}
