% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_temp.R
\name{remove_temp}
\alias{remove_temp}
\title{Remove Temporary Variables from a Dataset}
\usage{
remove_temp(.data)
}
\arguments{
\item{.data}{A data frame or data frame extension}
}
\value{
The data frame with temporary variables removed
}
\description{
This packages uses the convention of naming temporary variables with
the pattern \verb{.\{semantic_name\}_tmp_}. This function removes all variables with
such a name.
}
