% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-states.R
\docType{data}
\name{states}
\alias{states}
\title{State Abbreviations and FIPS Codes}
\format{
A \code{tibble} with 57 rows and 2 columns:
\describe{
\item{state}{Character. Two-letter abbreviations for each state.}
\item{fips}{Character. Two-digit FIPS code for each state.}
}
}
\source{
This is a subset of the
\href{https://walker-data.com/tidycensus/reference/fips_codes.html}{fips_codes}
dataset in the tidycensus package
}
\usage{
states
}
\description{
\code{states} contains two letter state abbreviations and the matching FIPS code
for each state.
}
\keyword{datasets}
