% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_addr.R
\name{std_addr}
\alias{std_addr}
\title{Standardize Addresses}
\usage{
std_addr(x)
}
\arguments{
\item{x}{Character-like vector. This will be converted to character before
standardizing.}
}
\value{
A character vector with standardized addresses
}
\description{
\code{std_addr()} transliterates all characters to ASCII, replaces all symbols
except dashes (\code{-}) and apostrophes (\verb{'}) with a space, removes apostrophes,
removes extraneous whitespace, and converts to title case.
}
