% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{std_names}
\alias{std_names}
\title{Re-Define \code{std_names()} to Remove Trailing Initials}
\usage{
std_names(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
A standardized character vector
}
\description{
\code{std_names()} uses \code{coviData::std_names()} to parse \code{string}, then removes
trailing single letters preceded by whitespace.
}
