% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_phone.R
\name{std_phone}
\alias{std_phone}
\title{Standardize Phone Numbers}
\usage{
std_phone(x, dialr = test_dialr(), quiet = TRUE)
}
\arguments{
\item{x}{A vector convertible to \code{character} format containing phone numbers}

\item{dialr}{Should the \code{dialr} package be used to check for valid number
ranges? The default uses dialr if it is installed and working.}

\item{quiet}{Should the progress bar from dialr be shown? Ignored if
\code{dialr = FALSE}.}
}
\value{
A \code{character} vector of standardized phone numbers
}
\description{
\code{std_phone()} standardizes North American phone numbers to a 10-digit
national format. Invalid or non-North American numbers are converted to \code{NA}.
Note that \code{std_phone()} only check for truly valid number ranges when
\code{dialr = TRUE}; otherwise, all 10-digit numbers are considered valid.
}
