% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_phone.R
\name{test_dialr}
\alias{test_dialr}
\title{Test that the dialr Package Works}
\usage{
test_dialr(quiet = FALSE)
}
\arguments{
\item{quiet}{Should warnings be shown if dialr does not work? These will only
be shown once per session.}
}
\value{
A boolean indicating whether dialr is functioning properly
}
\description{
dialr depends on the libphonenumber Java library and requires a working
Java Development Kit installation with environment variables set
appropriately. This function checks that dialr works and attempts to provide
informative messages about why it may not.
}
