% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_new_positive.R
\name{transform_new_positive}
\alias{transform_new_positive}
\title{Transform NBS Data to Format Necessary for Exporting to ACNS File}
\usage{
transform_new_positive(.data, date = attr(.data, "date"))
}
\arguments{
\item{.data}{Data from NBS snapshot file}

\item{date}{The date to use when creating the \code{DATE_ADDED} variable}
}
\value{
A \code{date_tbl}, which is a \code{tibble} with a \code{date} attribute
}
\description{
\code{transform_new_positive()} transforms data from NBS snapshot files to the
format needed for exports to the ACNS process.
}
