% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_acns_upload.R
\name{translate_acns_upload}
\alias{translate_acns_upload}
\title{Translate ACNS Data into Format Upload-able to the ACNS System}
\usage{
translate_acns_upload(.acns)
}
\arguments{
\item{.acns}{Prepared ACNS data}
}
\value{
ACNS data translated into the format needed by the ACNS system
}
\description{
Translate ACNS Data into Format Upload-able to the ACNS System
}
