% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_positive.R
\name{translate_positive}
\alias{translate_positive}
\title{Translate Positive Case Variables to ACNS}
\usage{
translate_positive(.data)
}
\arguments{
\item{.data}{A data frame containing positive NBS data (as \code{character})}
}
\value{
The input data with columns mapped to their ACNS counterparts
}
\description{
\code{translate_positive()} replaces columns from positive NBS data with their
equivalents in the ACNS file and drops any un-mapped columns.
This is not always a 1-to-1 mapping.
}
\details{
This function is memoised; it will cache return values and load them rather
than re-processing the same input data.
}
