% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_acns.R
\name{upload_acns}
\alias{upload_acns}
\title{Upload ACNS File to SFTP Server}
\usage{
upload_acns(
  .data,
  path = "ACNS/ACNS_NBS.csv",
  usr = Sys.getenv("sftp_usr"),
  pwd = Sys.getenv("sftp_pwd")
)
}
\arguments{
\item{.data}{ACNS data in a data frame or data frame extension}

\item{path}{The location to save the file on the SFTP server}

\item{usr}{The username for the SFTP server}

\item{pwd}{The password for the SFTP server}
}
\value{
The input data (invisibly)
}
\description{
\code{upload_acns()} validates and uploads ACNS data to the Shelby County SFTP
server.
}
