% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_positive.R
\name{upload_positive}
\alias{upload_positive}
\title{Upload Positives in NBS to ACNS Data}
\usage{
upload_positive(
  .data = prep_positive(),
  path = "nbs_cumulative.xlsx",
  usr = Sys.getenv("acns_usr"),
  pwd = Sys.getenv("acns_pwd")
)
}
\arguments{
\item{.data}{A \code{tibble} from \code{prep_positive()}}

\item{path}{The location to save the data on the SFTP server}

\item{usr}{The username for the SFTP server}

\item{pwd}{The password for the SFTP server}
}
\value{
\code{.data} (invisibly)
}
\description{
\code{upload_positive()} uploads data from \code{prep_positive()} to the ACNS SFTP
site for inclusion in the following day's ACNS_DAILY_OUT file.
}
