% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-coviData.R
\name{utils-coviData}
\alias{utils-coviData}
\title{coviData}
\description{
The coviData package serves as a backend for many of the SCHD R packages
related to COVID-19, including this package. Some functions that would
otherwise be included in this package are loaded here.
}
