% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sms_data.R
\name{validate_sms_data}
\alias{validate_sms_data}
\title{Validate Data for Inclusion in the ACNS File}
\usage{
validate_sms_data(.data, type = c("positive", "acns"))
}
\arguments{
\item{.data}{Data to validate}

\item{type}{Type of data; either \code{"positive"} or \code{"acns"} data}
}
\value{
The input data if validation passes, otherwise throws an error
}
\description{
Validate Data for Inclusion in the ACNS File
}
