% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_acns.R
\name{vroom_to_readxl}
\alias{vroom_to_readxl}
\title{Convert readr/vroom Column Specification to readxl Format}
\usage{
vroom_to_readxl(col_types)
}
\arguments{
\item{col_types}{A \code{col_types} specification for vroom/readr}
}
\value{
A \code{col_types} specification for readxl
}
\description{
Convert readr/vroom Column Specification to readxl Format
}
