% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_acns.R
\name{prep_acns}
\alias{prep_acns}
\title{Prepare ACNS Data for Use in Case Assignment}
\usage{
prep_acns(
  .data = download_acns(),
  incl_positive = TRUE,
  filter_acns = TRUE,
  filter_lab = TRUE,
  assign = FALSE,
  date = attr(.data, "date")
)
}
\arguments{
\item{.data}{Data frame or data frame extension containing ACNS data. If none
is supplied, the latest data will be downloaded.}

\item{incl_positive}{Should the output include new positives from NBS?}

\item{filter_acns}{Should duplicated data from the previous ACNS files be
filtered out?}

\item{filter_lab}{Should labs in the current ACNS data be filtered from NBS?}

\item{assign}{Is this data being used for assignment purposes (\code{TRUE}) or
sms notification (\code{FALSE})? The default is \code{FALSE}.}

\item{date}{The date that the data was reported. This is extracted from the
\code{date} attribute of a \code{date_tbl}.}
}
\value{
Prepared data in \code{tibble} format
}
\description{
\code{prep_acns()} prepares ACNS data for use in case assignment and NBS data
management processes. In addition to standardizing the dataset, it adds 3
variables:
\describe{
\item{school_age}{A logical indicating whether a case is school-aged}
\item{long_term_care}{A logical indicating whether a case is housed in a
long-term care facility}
\item{duplicate}{A logical indicating whether a case is a duplicate}
}
}
