% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_positive.R
\name{prep_positive}
\alias{prep_positive}
\title{Transform NBS Data to Format Necessary for Exporting to ACNS File}
\usage{
prep_positive(
  .data = load_positive(),
  date = attr(.data, "date"),
  phone = c("cell", "home", "work")
)
}
\arguments{
\item{.data}{Data from NBS snapshot file}

\item{date}{The date to use when creating the \code{DATE_ADDED} variable}
}
\value{
A \code{date_tbl}, which is a \code{tibble} with a \code{date} attribute
}
\description{
\code{prep_positive()} transforms data from NBS snapshot files to the
format needed for exports to the ACNS process.
}
