% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_phone.R
\name{std_phone}
\alias{std_phone}
\title{Standardize Phone Numbers}
\usage{
std_phone(x)
}
\arguments{
\item{x}{A vector convertible to \code{character} format containing phone numbers}
}
\value{
A \code{character} vector of standardized phone numbers
}
\description{
\code{std_phone()} standardizes North American phone numbers to a 10-digit
national format. Invalid or non-North American numbers are converted to \code{NA}.
}
